requirements.txt
feedparser==6.0.10
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
sqlalchemy==2.0.28
alembic==1.13.1 